[CRVA-3] Add WPF pages support and project refactored to valid template
