+++ .gitignore
@@ -1,0 +1,8 @@
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+
+# dependencies
+/Kalkulator/bin
+/Kalkulator/obj
+/CarRentalVoiceAssistant/bin
+/CarRentalVoiceAssistant/obj
+/.vs
