+++ CarRentalVoiceAssistant/.vs/ProjectSettings.json
@@ -1,0 +1,3 @@
+{
+  "CurrentProjectSetting": null
+}
