+++ CarRentalVoiceAssistant/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="Kalkulator.App"
+﻿<Application x:Class="CarRentalVoiceAssistant.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:Kalkulator"
+             xmlns:local="clr-namespace:CarRentalVoiceAssistant"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
