+++ CarRentalVoiceAssistant/App.xaml.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows;
 
-namespace Kalkulator
+namespace CarRentalVoiceAssistant
 {
     /// <summary>
     /// Logika interakcji dla klasy App.xaml
