+++ CarRentalVoiceAssistant/CarRentalVoiceAssistant.csproj
@@ -6,8 +6,8 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{0F8C859A-BE8C-4033-880C-D4D49239D260}</ProjectGuid>
     <OutputType>WinExe</OutputType>
-    <RootNamespace>Kalkulator</RootNamespace>
-    <AssemblyName>Kalkulator</AssemblyName>
+    <RootNamespace>CarRentalVoiceAssistant</RootNamespace>
+    <AssemblyName>CarRentalVoiceAssistant</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{60dc8134-eba5-43b8-bcc9-bb4bc16c2548};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
@@ -81,6 +81,16 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
+    <Compile Include="carChoose.xaml.cs">
+      <DependentUpon>carChoose.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="welcomeScreen.xaml.cs">
+      <DependentUpon>welcomeScreen.xaml</DependentUpon>
+    </Compile>
+    <Page Include="carChoose.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
     <Page Include="MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
@@ -93,6 +103,10 @@
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
+    <Page Include="welcomeScreen.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs">
