+++ CarRentalVoiceAssistant/MainWindow.xaml
@@ -1,0 +1,12 @@
+﻿<Window x:Class="CarRentalVoiceAssistant.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:CarRentalVoiceAssistant"
+        mc:Ignorable="d"
+        Title="Asystent Magierra" Height="352" Width="404.289">
+    <Grid>
+        <Frame Name="Main"/>
+    </Grid>
+</Window>
