+++ CarRentalVoiceAssistant/Properties/Resources.resx
