+++ CarRentalVoiceAssistant/carChoose.xaml
@@ -1,0 +1,15 @@
+﻿<Page x:Class="CarRentalVoiceAssistant.carChoose"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:CarRentalVoiceAssistant"
+      mc:Ignorable="d" 
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="carChoose">
+
+    <Grid>
+        <Label Content="Wybierz pojazd" HorizontalAlignment="Left" Margin="231,111,0,0" VerticalAlignment="Top" Width="384" Height="104" FontSize="36"/>
+
+    </Grid>
+</Page>
