+++ CarRentalVoiceAssistant/welcomeScreen.xaml
@@ -1,0 +1,15 @@
+﻿<Page x:Class="CarRentalVoiceAssistant.welcomeScreen"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:CarRentalVoiceAssistant"
+      mc:Ignorable="d" 
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="Magierra Assistant - Welcome">
+
+    <Grid>
+        <Label Content="Witaj w asystencie" HorizontalAlignment="Left" Margin="64,72,0,0" VerticalAlignment="Top" Height="57" Width="144" FontSize="16"/>
+        <Button Content="Button" HorizontalAlignment="Left" Margin="83,134,0,0" VerticalAlignment="Top" Width="75" Click="Button_Click"/>
+    </Grid>
+</Page>
