+++ CarRentalVoiceAssistant/welcomeScreen.xaml.cs
@@ -1,0 +1,33 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
+
+namespace CarRentalVoiceAssistant
+{
+    /// <summary>
+    /// Logika interakcji dla klasy welcomeScreen.xaml
+    /// </summary>
+    public partial class welcomeScreen : Page
+    {
+        public welcomeScreen()
+        {
+            InitializeComponent();
+        }
+
+        private void Button_Click(object sender, RoutedEventArgs e)
+        {
+
+        }
+    }
+}
