--- Kalkulator/MainWindow.xaml
@@ -1,36 +1,0 @@
-﻿<Window x:Class="Kalkulator.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Kalkulator"
-        mc:Ignorable="d"
-        Title="Kalkulator" Height="352" Width="404.289">
-    <Grid Margin="0,0,40,0">
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="0*"/>
-            <ColumnDefinition Width="119*"/>
-            <ColumnDefinition Width="237*"/>
-        </Grid.ColumnDefinitions>
-        <TextBox Name="Obliczenia" Grid.ColumnSpan="3" HorizontalAlignment="Left" Height="23" Margin="44,29,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="256"/>
-        <Button Name="B1" Content="1" HorizontalAlignment="Left" Margin="71,104,0,0" VerticalAlignment="Top" Width="37" Height="33" Click="Button_Click_1" Grid.ColumnSpan="2"/>
-        <Button Name="B2" Content="2" HorizontalAlignment="Left" Margin="1,104,0,0" VerticalAlignment="Top" Width="37" Height="33" Click="Button_Click_2" Grid.Column="2"/>
-        <Button Name="B3" Content="3" HorizontalAlignment="Left" Margin="49,104,0,0" VerticalAlignment="Top" Width="37" Height="33" Click="Button_Click_3" Grid.Column="2"/>
-        <Button Name="B4" Content="4" HorizontalAlignment="Left" Margin="71,150,0,0" VerticalAlignment="Top" Width="37" Height="33" Click="Button_Click_4" Grid.Column="1"/>
-        <Button Name="B5" Content="5" HorizontalAlignment="Left" Margin="0,151,0,0" VerticalAlignment="Top" Width="37" Height="33" Click="Button_Click_5" Grid.Column="2"/>
-        <Button Name="B6" Content="6" HorizontalAlignment="Left" Margin="49,152,0,0" VerticalAlignment="Top" Width="37" Height="33" Click="Button_Click_6" Grid.Column="2"/>
-        <Button Name="B7" Content="7" HorizontalAlignment="Left" Margin="72,193,0,0" VerticalAlignment="Top" Width="37" Height="33" Click="Button_Click_7" Grid.Column="1"/>
-        <Button Name="B8" Content="8" HorizontalAlignment="Left" Margin="0,194,0,0" VerticalAlignment="Top" Width="37" Height="33" Click="Button_Click_8" Grid.Column="2"/>
-        <Button Name="B9" Content="9" HorizontalAlignment="Left" Margin="50,195,0,0" VerticalAlignment="Top" Width="37" Height="33" Click="Button_Click_9" Grid.Column="2"/>
-        <Button Name="BPlus" Content="+" HorizontalAlignment="Left" Margin="115,103,0,0" VerticalAlignment="Top" Width="37" Height="33" Click="Button_Click_Plus" Grid.Column="2"/>
-        <Button Name="BMinus" Content="-" HorizontalAlignment="Left" Margin="115,153,0,0" VerticalAlignment="Top" Width="37" Height="33" Click="Button_Click_Minus" Grid.Column="2"/>
-        <Button Name="BMultiply" Content="*" HorizontalAlignment="Left" Margin="164,104,0,0" VerticalAlignment="Top" Width="37" Height="33" Click="Button_Click_Multiply" Grid.Column="2"/>
-        <Button Name="BEquality" Content="=" HorizontalAlignment="Left" Margin="167,200,0,0" VerticalAlignment="Top" Width="37" Height="33" Click="Button_Click_Equality" Grid.Column="2"/>
-        <Button Name="B0" Content="0" HorizontalAlignment="Left" Margin="0,237,0,0" VerticalAlignment="Top" Width="37" Height="33" Click="Button_Click_0" Grid.Column="2"/>
-        <Button Name="BDivide" Content="/" HorizontalAlignment="Left" Margin="165,154,0,0" VerticalAlignment="Top" Width="37" Height="33" Click="Button_Click_Divide" Grid.Column="2"/>
-        <Button Content="C" HorizontalAlignment="Left" Margin="113,199,0,0" VerticalAlignment="Top" Width="37" Height="33" Click="Button_Click_Clear" Grid.Column="2"/>
-        <Label Name="Liczba1" Grid.ColumnSpan="3" Content="Liczba1&#xD;&#xA;" HorizontalAlignment="Left" Margin="42,65,0,0" VerticalAlignment="Top" Width="95"/>
-        <Label Name="Liczba2" Content="Liczba2" Grid.Column="2" HorizontalAlignment="Left" Margin="66,65,0,0" VerticalAlignment="Top" Width="80"/>
-
-    </Grid>
-</Window>
